refactor: split layer.py into multiple files for avoiding circular imports
