+++ mapproxy/source/ogcapitiles.py
@@ -20,10 +20,11 @@ import sys
 from threading import Lock
 from typing import Optional
 
+from mapproxy.layer.map_layer import MapLayer
 from mapproxy.client.http import HTTPClientError
 from mapproxy.grid.tile_grid import tile_grid_from_ogc_tile_matrix_set
 from mapproxy.image.opts import ImageOptions
-from mapproxy.layer import BlankImage, MapLayer
+from mapproxy.layer import BlankImage
 from mapproxy.extent import MapExtent, DefaultMapExtent
 from mapproxy.source import SourceError, InvalidSourceQuery
 from mapproxy.srs import ogc_crs_url_to_auth_code
